move `PyDatafusionTable` to `catalog` and split moar stuff
