+++ rerun_py/src/catalog/task.rs
@@ -8,8 +8,7 @@ use pyo3::{Py, PyRef, PyResult, Python, pyclass, pymethods};
 use re_log_types::hash::Hash64;
 use re_protos::common::v1alpha1::TaskId;
 
-use crate::catalog::{PyCatalogClientInternal, to_py_err};
-use crate::dataframe::PyDataFusionTable;
+use super::{PyCatalogClientInternal, PyDataFusionTable, to_py_err};
 
 /// A handle on a remote task.
 #[pyclass(name = "Task")]
